refactor: split failure.rs into modular structure
Split failure.rs into logical modules following Rust 2018 patterns:
- failure/metadata.rs - FailureMetadata definition
- failure/escalation_stage.rs - EscalationStage enum and logic
- failure/failure_info.rs - FailureInfo implementation
- failure/failure_event.rs - FailureEvent enum
- failure/tests.rs - Test module

Also cleaned up unused imports in actor/system.rs and mailbox.rs.

All 131 tests passing
