+++ modules/actor-core/src/failure/failure_event.rs
@@ -1,0 +1,6 @@
+use super::FailureInfo;
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum FailureEvent {
+  RootEscalated(FailureInfo),
+}
