+++ modules/actor-core/src/mailbox.rs
@@ -2,7 +2,7 @@ mod messages;
 mod queue_mailbox;
 mod traits;
 
-pub use messages::{PriorityChannel, PriorityEnvelope, SystemMessage};
+pub use messages::{PriorityEnvelope, SystemMessage};
 pub use queue_mailbox::{MailboxOptions, QueueMailbox, QueueMailboxProducer, QueueMailboxRecv};
 pub use traits::{Mailbox, MailboxPair, MailboxRuntime, MailboxSignal};
